rename pr-cd
