+++ .github/workflows/cd.yml
@@ -66,7 +66,7 @@ jobs:
           result-encoding: string
           script: |
             const cwd = '${{ github.workspace }}';
-            const { parseBranchConfig } = await import(`${cwd}/packages/scripts/src/pr-cd.mjs`);
+            const { parseBranchConfig } = await import(`${cwd}/packages/scripts/src/gha-pr-helpers.mjs`);
             const ref = parseBranchConfig(context);
             console.log(ref);
             return ref;
@@ -88,7 +88,7 @@ jobs:
           result-encoding: string
           script: |
             const cwd = '${{ github.workspace }}';
-            const { findControlText } = await import(`${cwd}/packages/scripts/src/pr-cd.mjs`);
+            const { findControlText } = await import(`${cwd}/packages/scripts/src/gha-pr-helpers.mjs`);
             const control = await findControlText(context, github);
             console.log('Control text length:', control?.body?.length);
             return JSON.stringify(control ?? {});
@@ -102,7 +102,7 @@ jobs:
         with:
           script: |
             const cwd = '${{ github.workspace }}';
-            const { parseDeployConfig } = await import(`${cwd}/packages/scripts/src/pr-cd.mjs`);
+            const { parseDeployConfig } = await import(`${cwd}/packages/scripts/src/gha-pr-helpers.mjs`);
 
             const allDeploys = parseDeployConfig({
               ...${{ steps.control.outputs.result }},
