+++ .github/workflows/cleanup-auto-deploy.yml
@@ -57,7 +57,7 @@ jobs:
         with:
           script: |
             const cwd = '${{ github.workspace }}';
-            const { findDeploysToCleanUp } = await import(`${cwd}/packages/scripts/src/pr-cd.mjs`);
+            const { findDeploysToCleanUp } = await import(`${cwd}/packages/scripts/src/gha-pr-helpers.mjs`);
 
             const controls = ${{ toJson(steps.list.outputs.list) }};
             const down = await findDeploysToCleanUp(controls, ${{ github.event.inputs.ttl || 24 }}, context, github);
