+++ packages/scripts/tests/gha-pr-helpers.test.mjs
@@ -1,5 +1,5 @@
 import test from 'tape';
-import { parseDeployConfig } from '../src/pr-cd.mjs';
+import { parseDeployConfig } from '../src/gha-pr-helpers.mjs';
 
 function withoutOptions({ ...deploy }) {
   delete deploy.options;
